requirements.txt
asgiref==3.7.2
Django==5.0.1
django-autoslug==1.9.9
sqlparse==0.4.4
